backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["el-collins <charlescollins476@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
passlib = "^1.7.4"
httpx = "^0.27.0"
uvicorn = "^0.29.0"
pydantic = {extras = ["email"], version = "^2.6.4"}
motor = "^3.4.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
